Made multitasking more better and added Thread attributes =D

+++ src/Kernel/Kernel_alpha/Thread.cs
@@ -58,9 +58,9 @@ namespace Kernel_alpha
             IRQ.Timer();
         }
 
-        public static void Sleep(int Cycles)
+        public static void Sleep(uint Cycles)
         {
-            MT.Tasks[MT.CurrentTask].state = Cycles;//Cycles should be positive =P
+            MT.Tasks[MT.CurrentTask].state = (int)Cycles;//Cycles should be positive =P
             IRQ.Timer();
         }
     }
